WT-14572 MacOS disagg build fixes (#11912)
Some stats were the wrong type, and whatever the compiler is on our
macOS machines didn't like the stat increment code. Rather than change
all of that, just convert the underlying stat type to match all the
others.

+++ src/include/btmem.h
@@ -430,9 +430,9 @@ struct __wt_page_modify {
     wt_shared uint64_t update_txn;
 
     /* Dirty bytes added to the cache. */
-    wt_shared size_t bytes_dirty;
-    wt_shared size_t bytes_updates;
-    wt_shared size_t bytes_delta_updates;
+    wt_shared uint64_t bytes_dirty;
+    wt_shared uint64_t bytes_updates;
+    wt_shared uint64_t bytes_delta_updates;
 
     /*
      * When pages are reconciled, the result is one or more replacement blocks. A replacement block
